+++ CHANGELOG.md
@@ -1,7 +1,4 @@
 ## 1.0.0
-* The first version
-
-## 1.0.1
 * Update iOS SDK (to 0.0.15)
 * Update Android SDK (to 0.0.19)
 * handle Pagecall events (onLoaded, onMessage, onTerminated)
