+++ android/settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'pagecall_flutter'
+rootProject.name = 'flutter_pagecall'
