+++ example/pubspec.yaml
@@ -1,5 +1,5 @@
-name: pagecall_flutter_example
-description: Demonstrates how to use the pagecall_flutter plugin.
+name: flutter_pagecall_example
+description: Demonstrates how to use the flutter_pagecall plugin.
 # The following line prevents the package from being accidentally published to
 # pub.dev using `flutter pub publish`. This is preferred for private packages.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
@@ -17,9 +17,9 @@ dependencies:
   flutter:
     sdk: flutter
 
-  pagecall_flutter:
+  flutter_pagecall:
     # When depending on this package from a real application you should use:
-    #   pagecall_flutter: ^x.y.z
+    #   flutter_pagecall: ^x.y.z
     # See https://dart.dev/tools/pub/dependencies#version-constraints
     # The example app is bundled with the plugin so we use a path dependency on
     # the parent directory to use the current plugin's version.
