+++ ios/flutter_pagecall.podspec
@@ -1,9 +1,9 @@
 #
 # To learn more about a Podspec see http://guides.cocoapods.org/syntax/podspec.html.
-# Run `pod lib lint pagecall_flutter.podspec` to validate before publishing.
+# Run `pod lib lint flutter_pagecall.podspec` to validate before publishing.
 #
 Pod::Spec.new do |s|
-  s.name             = 'pagecall_flutter'
+  s.name             = 'flutter_pagecall'
   s.version          = '0.0.1'
   s.summary          = 'A new Flutter plugin project.'
   s.description      = <<-DESC
