+++ lib/src/android_pagecallview.dart
@@ -3,8 +3,8 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
-import 'package:pagecall_flutter/pagecall_flutter.dart';
-import 'package:pagecall_flutter/platform_interface.dart';
+import 'package:flutter_pagecall/flutter_pagecall.dart';
+import 'package:flutter_pagecall/platform_interface.dart';
 
 class AndroidPagecallView extends PlatformPagecallView {
   @override
